Converting the application to use cobra
